Creo la clase FormatArtistInfoHelper
