+++ app/src/main/java/ayds/lisboa/songinfo/moredetails/model/repository/external/lastFM/LastFMInjector.kt
@@ -3,4 +3,7 @@ package ayds.lisboa.songinfo.moredetails.model.repository.external.lastFM
 object LastFMInjector {
 
     val lastFMService: LastFMService = LastFMServiceImpl()
+
+    fun getFormatArtistInfoHelper(name: String) : FormatArtistInfoHelper =
+        FormatArtistInfoHelperImpl(name)
 }
